Moved ReactDOMRun, types.ts, and root.tsx, and update imports
