+++ frontend-streamhub-main/src/components/Navbar.tsx
@@ -47,11 +47,7 @@ const Navbar = () => {
               <Link to={`/update-profile`}>Update Profile</Link>
             </Button>
           </NavigationMenuItem>
-          <NavigationMenuItem>
-            <Button variant="ghost" className={buttonTextFormat} asChild>
-              <Link to={`/register`}>Register Account</Link>
-            </Button>
-          </NavigationMenuItem>
+
           <NavigationMenuItem>
             <Button
               variant="destructive"
