+++ frontend-streamhub-main/src/pages/HomePage.tsx
@@ -1,16 +1,8 @@
-import { Link } from "react-router-dom";
-
-import { Button } from "@/components/shadcn/ui/button";
-
 const HomePage = () => {
   return (
     <div className="flex flex-col text-white">
       <p>Home Page Placeholder</p>
       <br />
-      {/* <Link to={`/watch/one-piece`}>Click to route to Series Page</Link>
-      <Link to={`/watch/one-piece/123`}>Click to route to Watch Page</Link>
-      <Link to={`/watch-party/1456`}>Click to route to Watch Party Page</Link>
-      <Link to={`/hahaha`}>Click to route to Error Page</Link> */}
     </div>
   );
 };
