+++ frontend-streamhub-main/src/pages/LoginPage.tsx
@@ -4,7 +4,7 @@ import { useAppContext } from "@/contexts/AppContext";
 import { useForm } from "react-hook-form";
 import { useMutation } from "react-query";
 import * as apiClient from "@/api-client";
-import { User } from "@/types";
+import { User } from "@/utils/types";
 
 export type LoginFormData = {
   username: string;
