+++ frontend-streamhub-main/src/routes/ReactDOMRun.tsx
@@ -6,9 +6,8 @@ import HomePage from "@/pages/HomePage.tsx";
 import SeriesPage from "@/pages/SeriesPage.tsx";
 import WatchPage from "@/pages/WatchPage.tsx";
 import WatchPartyPage from "@/pages/WatchPartyPage.tsx";
-// import Dashboard from "@/pages/Dashboard.tsx";
 import UpdateProfilePage from "@/pages/UpdateProfilePage.tsx";
-import LandingPage from "../pages/LandingPage";
+
 import RegisterPage from "@/pages/RegisterPage";
 
 interface ReactDOMRunProps {}
